requirements.txt
torch
torchvision
tensorboard
scipy==1.13.0
numpy==1.24.3
matplotlib==3.9.0
opencv_python==4.9.0.80
tqdm==4.66.4
Pillow==10.3.0
h5py==3.11.0
PyQt5==5.15.2